% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_file}
\alias{check_input_file}
\title{Helper to check if input is exactly as expected csv}
\usage{
check_input_file(input_dataframe)
}
\arguments{
\item{input_dataframe}{}
}
\value{
TRUE if all is well
}
\description{
Helper to check if input is exactly as expected csv
}
