% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_abundance.R
\name{convert_to_abundance}
\alias{convert_to_abundance}
\title{Convert a known input format to the individual based abundance format that iNEXT expects}
\usage{
convert_to_abundance(input_dataframe, assemblage)
}
\arguments{
\item{input_dataframe}{dataframe with exactly the same columns as \code{tblWarandepark.csv}}
}
\value{
a list in the format as is expected by iNEXT when datatype is set to abundance
}
\description{
iNEXT expects \code{abundance} based data to be a list, with one child per
treatment/assemblage where the length of every child is equal to the number
of species in that treatment, and the sum of every child equal to the total
number of observed individuals.
}
\examples{
\dontrun{convert_to_abundance(warande, MicroMacro)}
}
