% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_incidence_freq.R
\name{convert_to_incidence_freq}
\alias{convert_to_incidence_freq}
\title{Convert a known input format to the incidence frequency format that iNEXT expects}
\usage{
convert_to_incidence_freq(input_dataframe, assemblage)
}
\arguments{
\item{input_dataframe}{dataframe with exactly the same columns as \code{tblWarandepark.csv}}

\item{assemblage}{variable in \code{input_dataframe} to group by.
Every value in this column will result in a seperate curve in the iNEXT output.}
}
\value{
a list in the format as is expected by iNEXT when datatype is set to incidence_freq
}
\description{
iNEXT expects incidence frequency data to be a list, with one child per
assemblage. Every child of this list starts with an integer equal to the
total number of sampling units (days, plots, ...), followed by one element
per species equal to the number of sampling units this species was observed
in.
}
\examples{
\dontrun{convert_to_incidence_freq(warande, MicroMacro)}
}
