% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_assemblage.R
\name{pull_assemblage}
\alias{pull_assemblage}
\title{Function to learn how to use data masking}
\usage{
pull_assemblage(input_dataframe, assemblage = NULL)
}
\arguments{
\item{input_dataframe}{the input dataframe}

\item{assemblage}{the variable to group by}
}
\value{
assemblage values
}
\description{
Function to learn how to use data masking
}
