% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_butterfly.R
\name{rarefy_butterfly}
\alias{rarefy_butterfly}
\title{Calculate rarefaction curve of Butterfly data}
\usage{
rarefy_butterfly(variables)
}
\arguments{
\item{variables}{input variables (skeleton)}
}
\value{
iNext object of rarefied and extrapolated values
}
\description{
Calculate rarefaction curve of Butterfly data
}
