% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlindersRarefaction-package.R
\docType{package}
\name{vlindersRarefaction-package}
\alias{vlindersRarefaction}
\alias{vlindersRarefaction-package}
\title{vlindersRarefaction: Use iNEXT to create Rarefaction curves for citizen science Butterfly data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pietrh.github.io/vlindersRarefaction/}
  \item \url{https://github.com/PietrH/vlindersRarefaction}
  \item Report bugs at \url{https://github.com/PietrH/vlindersRarefaction/issues}
}

}
\author{
\strong{Maintainer}: Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID})

Other contributors:
\itemize{
  \item Dirk Maes \email{dirk.maes@inbo.be} (\href{https://orcid.org/0000-0002-7947-3788}{ORCID}) [contributor]
}

}
\keyword{internal}
