% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waarnemingen_dirk.R
\docType{data}
\name{waarnemingen_dirk}
\alias{waarnemingen_dirk}
\title{Example waarnemingen.be export}
\format{
A dataframe with some exported waarnemingen.be observations
}
\usage{
waarnemingen_dirk
}
\description{
Example waarnemingen.be export
}
\examples{
\dontrun{
 waaarnemingen_dirk
}
}
\keyword{datasets}
