% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warande.R
\docType{data}
\name{warande}
\alias{warande}
\title{Warande butterfly observations}
\format{
A dataframe with the expected structure to be converted in this package
}
\usage{
warande
}
\description{
Warande butterfly observations
}
\examples{
\dontrun{
 warande
}
}
\keyword{datasets}
